fix: request content update also when focused index is -1 (#3618)
