+++ packages/combo-box/src/vaadin-combo-box-scroller.js
@@ -250,9 +250,16 @@ export class ComboBoxScroller extends PolymerElement {
   }
 
   /** @private */
-  __focusedIndexChanged(index) {
-    if (this.__virtualizer && index >= 0) {
+  __focusedIndexChanged(index, oldIndex) {
+    if (!this.__virtualizer) {
+      return;
+    }
+
+    if (index !== oldIndex) {
       this.requestContentUpdate();
+    }
+
+    if (index >= 0) {
       this.scrollIntoView(index);
     }
   }
